Add files via upload
Feedback Recieved

The project adopted a systematic approach to evaluating the utility of co-pilot for software translation. It was well structured and evaluated simple and complex programmes and functions for typed and untyped languages. The evaluation and conclusions highlighted and discussed the weaknesses of LLMs in code translation. The presentation was very clear and well organised. === The project involves evaluating CoPilot's translation of two speech quality models from a source language to a target language. Federico used a Likert scale to evaluate several dimensions of CoPilot's effectiveness, such as code preparation, complexity handling, developer satisfaction, and code quality. The project findings include that while CoPilot is effective for simple code translations involving classical algorithms, it lacks maturity in comprehensively automating the translation process. Federico's evaluation of CoPilot is comprehensive, using both simple benchmarks and complex real-life codes. His work is complex and original since no prior art is reported in this area. His methodology is sound and reproducible, and his report is well-structured, effectively presenting the technical complexity and findings of his work. I found Federico highly motivated to understand the buzz behind large language models (LLMs), and he has reasonably fulfilled the expectations in this project.